requirements.txt
streamlit
qrcode[pil]
openpyxl
easyocr
opencv-python-headless
pandas
fpdf
pillow
plotly